fix: test with debug output of outDir
